requirements.txt
typer[all]
inquirer
ujson
pytest
moviepy
Pillow